+++ CHANGELOG.md
@@ -24,6 +24,7 @@
 * Map savefiles in memory so that they are saved in states
 * Ram Watch can use pointers to the stack
 * Attempt to set tid of recovered threads using clone3
+* Track opened files that were deleted so we can map them in state loading
 
 ### Changed